requirements.txt
python-telegram-bot>=20.7
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
google-api-python-client>=2.120.0
google-auth>=2.29.0
